Lab 2 focuses on building a small project that incorporates object-oriented principles. In this lab, you are required to create three classes: Aims, Cart, and DigitalVideoDisc. The Aims class is responsible for executing the main method, while the Cart class is used to store ordered DVDs.
Lab Code:

The code for Lab 2 consists of the following classes:

Aims.java: This class contains the main method and serves as the entry point for the program. It manages the execution flow of the application.
Cart.java: The Cart class is designed to store information about the DVDs added to the shopping cart. It includes methods for adding, removing, and displaying the DVDs in the cart.
DigitalVideoDisc.java: This class represents a digital video disc (DVD) with properties such as title, category, director, length, and cost. It is used to create DVD objects that can be added to the cart.